# RestSwagger/swagger.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "RestSwagger",
    "version": "1.0"
  },
  "paths": {
    "/Cars": {
      "get": {
        "tags": [
          "Cars"
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Car"
                  }
                }
              },
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Car"
                  }
                }
              },
              "text/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Car"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Cars"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Car"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/Car"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/Car"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/Car"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Car"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/Car"
                }
              }
            }
          }
        }
      }
    },
    "/Cars/{id}": {
      "get": {
        "tags": [
          "Cars"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/Car"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Car"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/Car"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "Cars"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "properties": {
                  "Id": {
                    "type": "string",
                    "format": "uuid"
                  },
                  "StartRent": {
                    "type": "string",
                    "format": "date-time"
                  },
                  "EndRent": {
                    "type": "string",
                    "format": "date-time"
                  },
                  "IsRent": {
                    "type": "boolean"
                  }
                }
              },
              "encoding": {
                "Id": {
                  "style": "form"
                },
                "StartRent": {
                  "style": "form"
                },
                "EndRent": {
                  "style": "form"
                },
                "IsRent": {
                  "style": "form"
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/Car"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Car"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/Car"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Cars"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "Car": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "startRent": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "endRent": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "isRent": {
            "type": "boolean",
            "readOnly": true
          }
        },
        "additionalProperties": false
      }
    }
  }
}